% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QSM_functions.R
\name{checkQSM}
\alias{checkQSM}
\title{Check if QSM must be updated}
\usage{
checkQSM(qsm, precision = 2L)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{precision}{Number of decimal points to be compared.}
}
\value{
A logical. \code{TRUE} if the object is likely up-to-date, \code{FALSE} if
the object needs to be updated.
}
\description{
\code{checkQSM} conducts some superficial tests whether the \code{overview}
and \code{branch} data are up-to-date with the \code{cylinder} data of a
\code{QSM} object.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# check qsm
checkQSM(qsm)
}
\seealso{
\code{\link{updateQSM}}
}
