% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{conductive_paths}
\alias{conductive_paths}
\title{Get conductive paths of branch tips}
\usage{
conductive_paths(qsm)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}
}
\value{
A \code{list} containing the conductive path lengths and nodes.
}
\description{
\code{conductive_paths} identifies all branch tips and calculates the total
length, the length segments, the total number of nodes and the nodes taken
of all cylinders between branch tips and stem base.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# get conductive paths
conductive_paths(qsm)
}
