% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_location}
\alias{get_location}
\title{Get location of the stem base}
\usage{
get_location(qsm)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}
}
\value{
A named \code{numeric} containing the \code{xyz}-coordinates of the stem base.
}
\description{
\code{get_location} obtains the location of the stem base.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# get stem base location
stem_base <- get_location(qsm)
}
\seealso{
\code{\link{set_location}}
}
