% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classQSM.R
\name{plot}
\alias{plot}
\alias{plot,QSM,missing-method}
\title{Plot a QSM object}
\usage{
plot(x, y, ...)

\S4method{plot}{QSM,missing}(
  x,
  y,
  col = NULL,
  col_var = "BranchOrder",
  pal = grDevices::rainbow,
  sides = 6,
  lit = FALSE,
  center = TRUE
)
}
\arguments{
\item{x}{An object of class \code{QSM}.}

\item{y}{Not used.}

\item{col}{\code{character}, color used for coloring the whole
\code{QSM} in a single color, if trying to color according to a variable,
leave it empty.}

\item{col_var}{\code{character}, column name of the cylinder data that should
be used for coloring.}

\item{pal}{\code{function}, color palette used for coloring.}

\item{sides}{\code{integer}, number of sides in the polygon cross section.}

\item{lit}{\code{boolean}, whether the scene should be with or without light.}

\item{center}{\code{boolean}, whether all cylinders should be centered at
the stem base.}
}
\value{
\code{rgl} plot of a QSM.
}
\description{
Displays a 3D \code{rgl} plot of a \code{QSM} object. Can be manipulated
using the package \code{rgl}.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# plot qsm
plot(qsm, col = "salmon4")
plot(qsm, col_var = "PositionInBranch")
}
\seealso{
\code{\link{readQSM}}
}
