% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_branches}
\alias{pruning_branches}
\title{todo}
\usage{
pruning_branches(qsm, branch_id, remove = FALSE)
}
\arguments{
\item{todo}{\code{todo}, todo.}
}
\value{
\code{todo}, todo.
}
\description{
\code{todo} todo.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune specified branches, without removing
not_removed <- pruning_branches(qsm, branch_id = c(5,33), remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune specified branches, with removing
removed <- pruning_branches(qsm, branch_id = c(3,111,333), remove = TRUE)

# plot qsm
plot(removed)
}
\seealso{
\code{\link{todo}}
}
