% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_conventional}
\alias{pruning_conventional}
\title{todo}
\usage{
pruning_conventional(
  qsm,
  threshold_m = 3,
  method = c("length", "height"),
  remove = FALSE
)
}
\arguments{
\item{todo}{\code{todo}, todo.}
}
\value{
\code{todo}, todo.
}
\description{
\code{todo} todo.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune specified branches, without removing
not_removed <- pruning_conventional(qsm, threshold_m = 3, method = "length", remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune specified branches, with removing
removed <- pruning_conventional(qsm, threshold_m = 3, method = "length", remove = TRUE)

# plot qsm
plot(removed)
}
\seealso{
\code{\link{todo}}
}
