% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_cylinders}
\alias{pruning_cylinders}
\title{Pruning according to cylinder ID}
\usage{
pruning_cylinders(qsm, cyl_id, remove = FALSE)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{cyl_id}{\code{integer}, cylinder IDs at which the tree should be pruned.}

\item{remove}{\code{boolean}, whether the to be pruned cylinders should be
removed (\code{TRUE}) or labelled (\code{FALSE}).}
}
\value{
\code{QSM}, with removed or labelled cylinders (column \code{cylinder$prune}).
}
\description{
\code{pruning_cylinders} prunes the tree at the specified cylinders and also
removes subsequent cylinders.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune at specified cylinders, without removing
not_removed <- pruning_cylinders(qsm, cyl_id = c(33,55,77), remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune at specified cylinders, with removing
removed <- pruning_cylinders(qsm, cyl_id = c(33,55,77), remove = TRUE)

# plot qsm
plot(removed)
}
\seealso{
\code{\link{pruning_branches}}
}
