% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_selective}
\alias{pruning_selective}
\title{todo}
\usage{
pruning_selective(qsm, diameter_m = 0.03, angle_deg = 40, remove = FALSE)
}
\arguments{
\item{todo}{\code{todo}, todo.}
}
\value{
\code{todo}, todo.
}
\description{
\code{todo} todo.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune specified branches, without removing
not_removed <- pruning_selective(qsm, diameter_m = 0.03, angle_deg = 40, remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune specified branches, with removing
removed <- pruning_selective(qsm, diameter_m = 0.03, angle_deg = 40, remove = TRUE)

# plot qsm
plot(removed)

}
\references{
Springmann, S., Rogers, R., & Spiecker, H. (2011). Impact of artificial
pruning on growth and secondary shoot development of wild cherry
(\emph{Prunus avium} L.). Forest Ecology and Management, 261(3), 764-769.
doi: \href{https://doi.org/10.1016/j.foreco.2010.12.007}{10.1016/j.foreco.2010.12.007}
}
\seealso{
\code{\link{todo}}
}
