% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_tips}
\alias{pruning_tips}
\title{Pruning of tips at specified diameter}
\usage{
pruning_tips(
  qsm,
  diameter_m = 0.07,
  length_m = 0,
  num_childs = 0,
  remove = FALSE
)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{diameter_m}{\code{numeric}, branch diameter in meters below which
branch tips should be pruned.}

\item{length_m}{\code{numeric}, minimum length of consecutive cylinders below
the diameter threshold.}

\item{num_childs}{\code{integer}, minimum number of consecutive cylinders
below the diameter threshold.}

\item{remove}{\code{boolean}, whether the to be pruned cylinders should be
removed (\code{TRUE}) or labelled (\code{FALSE}).}
}
\value{
\code{QSM}, with removed or labelled cylinders (column \code{cylinder$prune}).
}
\description{
\code{pruning_tips} prunes the tips of all branches below a specified branch
diameter. A minimum length and/or number of consecutive cylinders falling
below the specified branch diameter can be specified. If a tip is reached
before the length and number threshold can be met, the cylinders are removed
regardless of these parameters.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune at 5cm branch diameter, without removing
not_removed <- pruning_tips(qsm, diameter_m = 0.05, remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune at 5cm branch diameter, with removing
removed <- pruning_tips(qsm, diameter_m = 0.05, remove = TRUE)

# plot qsm
plot(removed)
}
\seealso{
\code{\link{pruning_conventional}}, \code{\link{pruning_selective}},
\code{\link{pruning_whorlwise}}
}
