% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_whorlwise}
\alias{pruning_whorlwise}
\title{Whorlwise pruning to a specified number of whorls}
\usage{
pruning_whorlwise(qsm, whorl_m = 0.3, remaining_whorls = 5, remove = FALSE)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{whorl_m}{\code{numeric}, maximum whorl stem interval in meters.}

\item{remaining_whorls}{\code{integer}, number of whorls that should remain
at the top of the crown.}

\item{remove}{\code{boolean}, whether the to be pruned cylinders should be
removed (\code{TRUE}) or labelled (\code{FALSE}).}
}
\value{
\code{QSM}, with removed or labelled cylinders (column \code{"prune"}).
}
\description{
\code{pruning_whorlwise} prunes all first order and their subsequent
branches belonging to whorls below a specified number of whorls. Whorls are
determined automatically.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune specified branches, without removing
not_removed <- pruning_whorlwise(qsm, whorl_m = 0.3, remaining_whorls = 5, remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune specified branches, with removing
removed <- pruning_whorlwise(qsm, whorl_m = 0.3, remaining_whorls = 5, remove = TRUE)

# plot qsm
plot(removed)

}
\references{
Springmann, S., Rogers, R., & Spiecker, H. (2011). Impact of artificial
pruning on growth and secondary shoot development of wild cherry
(\emph{Prunus avium} L.). Forest Ecology and Management, 261(3), 764-769.
doi: \href{https://doi.org/10.1016/j.foreco.2010.12.007}{10.1016/j.foreco.2010.12.007}
}
\seealso{
\code{\link{pruning_conventional}}, \code{\link{pruning_selective}}
}
