% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateQSM.R
\name{updateQSM}
\alias{updateQSM}
\title{Update QSM}
\usage{
updateQSM(qsm, method = "TreeQSM")
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}
}
\value{
An updated object of class \code{QSM}.
}
\description{
\code{updateQSM} updates the \code{overview} and \code{branch} data of a
\code{QSM} object based on its \code{cylinder} data.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# delete some cylinders
qsm@cylinder <- qsm@cylinder[qsm@cylinder$BranchOrder <= 3,]

# update qsm
updateQSM(qsm)
}
\seealso{
\code{\link{checkQSM}}
}
