% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_branches}
\alias{pruning_branches}
\title{Pruning according to branch ID}
\usage{
pruning_branches(qsm, branch_id, remove = FALSE)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{branch_id}{\code{integer}, branch IDs of to be pruned branches.}

\item{remove}{\code{boolean}, whether the to be pruned cylinders should be
removed (\code{TRUE}) or labelled (\code{FALSE}).}
}
\value{
\code{QSM}, with removed or labelled cylinders (column \code{"prune"}).
}
\description{
\code{pruning_branches} prunes the tree at the specified branches and also
removes subsequent branches.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune specified branches, without removing
not_removed <- pruning_branches(qsm, branch_id = c(5,33), remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune specified branches, with removing
removed <- pruning_branches(qsm, branch_id = c(3,111,333), remove = TRUE)

# plot qsm
plot(removed)
}
\seealso{
\code{\link{pruning_cylinders}}
}
