% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_cylinders}
\alias{pruning_cylinders}
\title{todo}
\usage{
pruning_cylinders(qsm, cyl_id, remove = FALSE)
}
\arguments{
\item{todo}{\code{todo}, todo.}
}
\value{
\code{todo}, todo.
}
\description{
\code{todo} todo.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune at specified cylinders, without removing
not_removed <- pruning_cylinders(qsm, cyl_id = c(33,55,77), remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune at specified cylinders, with removing
removed <- pruning_cylinders(qsm, cyl_id = c(33,55,77), remove = TRUE)

# plot qsm
plot(removed)
}
\seealso{
\code{\link{todo}}
}
