% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruning.R
\name{pruning_selective}
\alias{pruning_selective}
\title{Selective pruning according to specified branch diameter and angle}
\usage{
pruning_selective(qsm, diameter_m = 0.03, angle_deg = 40, remove = FALSE)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{diameter_m}{\code{numeric}, branch diameter in meters above which all
first order branches should be removed.}

\item{angle_deg}{\code{numeric}, angle between stem and branches in degree
below which all first order branches should be removed.}

\item{remove}{\code{boolean}, whether the to be pruned cylinders should be
removed (\code{TRUE}) or labelled (\code{FALSE}).}
}
\value{
\code{QSM}, with removed or labelled cylinders (column \code{"prune"}).
}
\description{
\code{pruning_selective} prunes all first order and their subsequent
branches which have a larger starting diameter or a smaller branch angle than
specified.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# prune specified branches, without removing
not_removed <- pruning_selective(qsm, diameter_m = 0.03, angle_deg = 40, remove = FALSE)

# plot qsm
plot(not_removed, col_var = "prune")

# prune specified branches, with removing
removed <- pruning_selective(qsm, diameter_m = 0.03, angle_deg = 40, remove = TRUE)

# plot qsm
plot(removed)

}
\references{
Springmann, S., Rogers, R., & Spiecker, H. (2011). Impact of artificial
pruning on growth and secondary shoot development of wild cherry
(\emph{Prunus avium} L.). Forest Ecology and Management, 261(3), 764-769.
doi: \href{https://doi.org/10.1016/j.foreco.2010.12.007}{10.1016/j.foreco.2010.12.007}
}
\seealso{
\code{\link{pruning_conventional}}, \code{\link{pruning_whorlwise}}
}
