% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readQSM.R
\name{readQSM}
\alias{readQSM}
\title{Read QSM from Matlab file}
\usage{
readQSM(file_path, qsm_var = 1, qsm_idx = 1)
}
\arguments{
\item{file_path}{Path to the \code{*.mat} file.}

\item{qsm_var}{Name of the variable in the file in which the QSM is stored.
Uses per default the first variable in the file.}

\item{qsm_idx}{Index of the QSM in the variable is stored in.
Uses per default the first QSM in the variable.}
}
\value{
A new object of class \code{QSM}.
}
\description{
\code{readQSM} reads in one QSM from a \code{*.mat} file created with
\href{https://github.com/InverseTampere/TreeQSM}{TreeQSM} in Matlab.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# inspect qsm
print(qsm)
}
