% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_location}
\alias{set_location}
\title{Set location of the stem base}
\usage{
set_location(qsm, location)
}
\arguments{
\item{qsm}{An object of class \code{QSM}.}

\item{location}{Coordinates of a point which is supposed to be moved to
\code{(0|0|0)}. A list or a data.frame with items or columns \code{xyz}.}
}
\value{
An object of class \code{QSM}.
}
\description{
\code{set_location} changes the location of the QSM so that \code{location}
is moved to the point \code{(0|0|0)}.
}
\examples{
# load qsm
file_path <- system.file("extdata", "QSM_Juglans_regia_M.mat", package="qsm2r")
qsm <- readQSM(file_path)

# get stem base location
stem_base <- get_location(qsm)

# shift stem base to (0|0|0)
qsm <- set_location(qsm, stem_base)
}
\seealso{
\code{\link{get_location}}
}
